refactor(dto): cleanse package structure
- remove unnecessary request dto
