+++ src/main/java/chess/domain/game/Game.java
@@ -2,8 +2,8 @@ package chess.domain.game;
 
 import chess.domain.event.MoveCommand;
 import chess.domain.game.statistics.GameResult;
-import chess.dto.response.GameDto;
-import chess.dto.response.board.ConsoleBoardViewDto;
+import chess.dto.GameDto;
+import chess.dto.board.ConsoleBoardViewDto;
 
 public interface Game {
